added log

+++ Schoole/Services/ClassroomsService/DeleteClassroom.cs
@@ -1,16 +1,22 @@
 ﻿using Schoole.Interfaces;
+using Schoole.Models;
+using Schoole.Repositories.Database;
 
 namespace Schoole.Services.ClassroomsService
 {
     public interface IDeleteClassroom
     {
-        void Execute(int classroomId);
+        Task Execute(int classroomId);
     }
-    public class DeleteClassroom(IClassroomRepository classroomRepository) : IDeleteClassroom
+    public class DeleteClassroom(IClassroomRepository classroomRepository, ILogService logService) : IDeleteClassroom
     {
-        public void Execute(int classroomId)
+        public async Task Execute(int classroomId)
         {
+            var classroom = classroomRepository.GetClassroomById(classroomId);
+            if (classroom == null) return;
+
             classroomRepository.Delete(classroomId);
+            await logService.LogInfo($"Classroom deleted: {classroom.Name}");
         }
     }
 }
